+++ tests/TestCase/Model/Entity/PostTest.php
@@ -26,7 +26,6 @@ use Cake\Cache\Cache;
 use Cake\ORM\TableRegistry;
 use Cake\TestSuite\TestCase;
 use MeCms\Model\Entity\Post;
-use MeTools\Utility\Youtube;
 
 /**
  * PostTest class
@@ -90,38 +89,10 @@ class PostTest extends TestCase
      */
     public function testNoAccessibleProperties()
     {
-        $entity = new Post();
+        $entity = new Post;
 
         $this->assertFalse($entity->accessible('id'));
         $this->assertFalse($entity->accessible('modified'));
-    }
-
-    /**
-     * Test for `_getPreview()` method
-     * @test
-     */
-    public function testPreviewGetMutator()
-    {
-        $entity = new Post();
-
-        $this->assertNull($entity->preview);
-
-        $entity->text = 'This is a simple text';
-        $this->assertFalse($entity->preview);
-
-        $entity->text = '<img src=\'image.jpg\' /> Image before text';
-        $this->assertEquals('image.jpg', $entity->preview);
-
-        $expected = Youtube::getPreview('videoID');
-
-        $entity->text = '[youtube]videoID[/youtube]';
-        $this->assertEquals($expected, $entity->preview);
-
-        $entity->text = '[youtube]videoID[/youtube]Text';
-        $this->assertEquals($expected, $entity->preview);
-
-        $entity->text = '[youtube]videoID[/youtube] Text';
-        $this->assertEquals($expected, $entity->preview);
     }
 
     /**